Delete unused configs.
These are not installed by the ad hoc install.sh script. They're not in
use on the machines I use daily, either, so they really have no value
here anymore. If I need them in the future, I'll recover them from the
history.

This is a preparatory step for adopting yadm to manage my dotfiles.