Push yadm alt down to subdir.
The yadm alt ("##os.Darwin") will work only if the directory doesn't already
exist. On Macs, obviously ~/Library already exists for a user, so the alt
doesn't get installed.

Putting the ##os.Darwin alt on .../Library/KeyBindings works because that
directory doesn't exist by default. When the yadm-based dotfiles repo is cloned
and installed, ~/Library/KeyBindings will become a symlink as expected. (Or, if
KeyBindings already exists from a manual change, it'll be skipped with a
warning.)